add uint8_t casts

+++ src/SparkFun_SHTC3.cpp
@@ -181,7 +181,7 @@ SHTC3_Status_TypeDef SHTC3::checkID()
 		return exitOp(retval, __FILE__, __LINE__);
 	}
 
-	numBytesRx = _wire->requestFrom(SHTC3_ADDR_7BIT, numBytesRequest);
+	numBytesRx = _wire->requestFrom((uint8_t)SHTC3_ADDR_7BIT, numBytesRequest);
 	if (numBytesRx != numBytesRequest)
 	{
 		exitOp(SHTC3_Status_Error, __FILE__, __LINE__);
@@ -311,7 +311,7 @@ SHTC3_Status_TypeDef SHTC3::update()
 	case SHTC3_CMD_CSE_RHF_LPM:
 	case SHTC3_CMD_CSE_TF_NPM:
 	case SHTC3_CMD_CSE_TF_LPM:	   // Address+read will yield an ACK and then clock stretching will occur
-		numBytesRx = _wire->requestFrom(SHTC3_ADDR_7BIT, numBytesRequest);
+		numBytesRx = _wire->requestFrom((uint8_t)SHTC3_ADDR_7BIT, numBytesRequest);
 		break;
 
 	case SHTC3_CMD_CSD_RHF_NPM:
